+++ src/rotsensor.cpp
@@ -1,7 +1,7 @@
 #include <Arduino.h>
 #include "rotsensor.hh"
 
-POTENTIO::POTENTIO(uint8_t _pin, unsigned bits, double range, double _scale):
+POTENTIO::POTENTIO(uint8_t _pin, unsigned bits, float range, float _scale):
     pin(_pin), res(range/pow(2,bits)), scale(_scale) {
         degree = analogRead(pin) * res * scale;
     }
@@ -11,21 +11,21 @@ int POTENTIO::getRawVal() {
     return rawVal;
 }
 
-double POTENTIO::getVoltVal() {
+float POTENTIO::getVoltVal() {
     getRawVal();
     volt = rawVal * res;
     return volt; 
 }
 
-double POTENTIO::getDegreeVal() {
+float POTENTIO::getDegreeVal() {
     getVoltVal();
     // low pass filter
-    static const double coef = 0.5;
+    static const float coef = 0.5;
     degree = coef*degree + (1.0 - coef)*(volt * scale);
 
     return degree;
 }
 
-void POTENTIO::setScale(double _scale) {
+void POTENTIO::setScale(float _scale) {
     scale = _scale;
 }
